+++ common/src/chain/mod.rs
@@ -17,6 +17,7 @@ pub mod block;
 pub mod config;
 pub mod gen_block;
 pub mod genesis;
+mod mlt;
 mod pow;
 pub mod tokens;
 pub mod transaction;
@@ -28,5 +29,6 @@ pub use block::Block;
 pub use config::ChainConfig;
 pub use gen_block::{GenBlock, GenBlockId};
 pub use genesis::Genesis;
+pub use mlt::Mlt;
 pub use pow::PoWChainConfig;
 pub use upgrades::*;
